Split device module into two modules
In the next few commits the device module will go away entirely,
while the devices module will take some method from the blockdevmgr
module.
